requirements.txt
pytorch
transformers
flwr
evaluate
datasets
pandas
sciket-learn